检查了菜单数据接口的名称。根据"注册控件"返回值现在决定要不要打开目录。

+++ examples/MFCApplication1/HttpUtil.cpp
@@ -164,7 +164,7 @@ namespace Alime
 		if (!connectedInternet) goto CLEANUP;
 
 		// open request
-		for (int i = 0; i < request.acceptTypes.size(); i++)
+		for (size_t i = 0; i < request.acceptTypes.size(); i++)
 		{
 			acceptTypes.push_back(request.acceptTypes.at(i).data());
 		}
